Moved `BasicScore` to production.
